+++ packages/workspace-fusion/src/lib/index.ts
@@ -1,4 +1,4 @@
-export * from './classes';
 export * from './types';
 export { useWorkspace } from './hooks';
-export { createFusionWorkspace } from './utils';
+export { Workspace } from './components';
+export type { WorkspaceProps } from './components';