+++ packages/workspace-fusion/src/lib/types/index.ts
@@ -1,5 +1,5 @@
 export * from './configuration';
-export * from './onClick';
+export * from './event';
 export * from './tabs';
 export * from './fusionBookmark';
 export * from './fusionController';