+++ packages/workspace-fusion/src/lib/utils/index.ts
@@ -6,6 +6,5 @@ export * from './sidesheet';
 export * from './statusBar';
 export * from './filter';
 export * from './addCustomTab';
-export * from './createFusionWorkspace';
 export * from './addConfig';
 export * from './viewController';